add drift type trait to use with consumers
